Move helper functions to new file and a bit of renaming
