+++ sdc/extensions/indexes/multi_index_ext.py
@@ -56,10 +56,21 @@ from sdc.hiframes.boxing import _infer_index_type, _unbox_index_data
 from sdc.extensions.indexes.indexes_generic import *
 
 from sdc.datatypes.indexes.multi_index_type import MultiIndexIteratorType
-from numba.core.extending import register_jitable
 from numba import literal_unroll
 from numba.typed import Dict, List
 from sdc.datatypes.sdc_typeref import MultiIndexTypeRef
+from sdc.extensions.indexes.multi_index_helpers import (
+        _multi_index_binop_helper,
+        _multi_index_append_level,
+        _multi_index_create_level,
+        _multi_index_create_levels_and_codes,
+        _multi_index_alloc_level_dict,
+        _multi_index_from_tuples_helper,
+        cat_array_equal,
+        next_codes_info,
+        next_codes_array,
+        factorize_level,
+    )
 
 
 @typeof_impl.register(pd.MultiIndex)
@@ -348,9 +359,15 @@ def pd_multi_index_overload(levels, codes, sortorder=None, names=None,
 
 
 @lower_builtin(MultiIndexTypeRef, types.VarArg(types.Any))
-def sdctyperef_call_impl(context, builder, sig, args):
+def multi_index_typeref_call_impl(context, builder, sig, args):
 
-    # FIXME: this hardcodes template number and selected dispatcher, refactor?
+    # FIXME_Numba#7111: this uses low-level API as a workaround for numba issue
+    # TO-DO: remove and use @overload(MultiIndexTypeRef), once issue is fixed
+    # and now we do the following:
+    # (1) lookup function type for the actual ctor (sdc_pandas_multi_index_ctor)
+    # (2) get compiled implementation for provided args (hardcodes 0 as selected overload template,
+    #     i.e. we rely on the fact that sdc_pandas_multi_index_ctor was overloaded only once)
+    # (3) get the function descriptor from compiled result and emit the call to it 
     call_sig = context.typing_context._resolve_user_function_type(
         sdc_pandas_multi_index_ctor,
         sig.args,
@@ -595,80 +612,6 @@ def pd_multi_index_names_overload(self):
     return pd_multi_index_names_impl
 
 
-# FIXME: move to a different file?
-def cat_array_equal(A, codes_A, B, codes_B):
-    pass
-
-
-@sdc_overload(cat_array_equal)
-def sdc_cat_array_equal_overload(A, codes_A, B, codes_B):
-
-    def sdc_cat_array_equal_impl(A, codes_A, B, codes_B):
-        if len(codes_A) != len(codes_B):
-            return False
-
-        # FIXME_Numba#5157: change to simple A == B when issue is resolved
-        eq_res_size = len(codes_A)
-        eq_res = np.empty(eq_res_size, dtype=types.bool_)
-        for i in numba.prange(eq_res_size):
-            eq_res[i] = A[codes_A[i]] == B[codes_B[i]]
-        return np.all(eq_res)
-
-    return sdc_cat_array_equal_impl
-
-
-@intrinsic
-def _multi_index_binop_helper(typingctx, self, other):
-    """ This function gets two multi_index objects each represented as
-    Tuple(levels) and Tuple(codes) and repacks these into Tuple of following
-    elements (self_level_0, self_codes_0, other_level_0, other_codes_0), etc
-    """
-
-    nlevels = len(self.levels)
-    if not len(self.levels) == len(other.levels):
-        assert True, "Cannot flatten MultiIndex of different nlevels"
-
-    elements_types = zip(self.levels, self.codes, other.levels, other.codes)
-    ret_type = types.Tuple([types.Tuple.from_types(x) for x in elements_types])
-
-    def codegen(context, builder, sig, args):
-        self_val, other_val = args
-
-        self_ctinfo = cgutils.create_struct_proxy(self)(
-                    context, builder, value=self_val)
-        self_levels = self_ctinfo.levels
-        self_codes = self_ctinfo.codes
-
-        other_ctinfo = cgutils.create_struct_proxy(other)(
-                    context, builder, value=other_val)
-        other_levels = other_ctinfo.levels
-        other_codes = other_ctinfo.codes
-
-        ret_tuples = []
-        for i in range(nlevels):
-            self_level_i = builder.extract_value(self_levels, i)
-            self_codes_i = builder.extract_value(self_codes, i)
-            other_level_i = builder.extract_value(other_levels, i)
-            other_codes_i = builder.extract_value(other_codes, i)
-
-            ret_tuples.append(
-                context.make_tuple(builder,
-                                   ret_type[i],
-                                   [self_level_i, self_codes_i, other_level_i, other_codes_i])
-            )
-
-            if context.enable_nrt:
-                context.nrt.incref(builder, ret_type[i][0], self_level_i)
-                context.nrt.incref(builder, ret_type[i][1], self_codes_i)
-                context.nrt.incref(builder, ret_type[i][2], other_level_i)
-                context.nrt.incref(builder, ret_type[i][3], other_codes_i)
-
-        res = context.make_tuple(builder, ret_type, ret_tuples)
-        return res
-
-    return ret_type(self, other), codegen
-
-
 @sdc_overload_method(MultiIndexType, 'equals')
 def pd_multi_index_equals_overload(self, other):
     if not isinstance(self, MultiIndexType):
@@ -701,33 +644,6 @@ def pd_multi_index_equals_overload(self, other):
     return pd_multi_index_equals_impl
 
 
-# FIXME: move to another file?
-def _build_index_map(self):
-    pass
-
-
-@sdc_overload(_build_index_map)
-def _build_index_map_ovld(self):
-
-    indexer_dtype = self.dtype
-    indexer_value_type = types.ListType(types.int64)
-
-    def _build_index_map(self):
-        indexer_map = Dict.empty(indexer_dtype, indexer_value_type)
-        for i in range(len(self)):
-            val = self[i]
-            index_list = indexer_map.get(val, None)
-            if index_list is None:
-                indexer_map[val] = List.empty_list(types.int64)
-                indexer_map[val].append(i)
-            else:
-                index_list.append(i)
-
-        return indexer_map
-
-    return _build_index_map
-
-
 @sdc_overload(operator.contains)
 def pd_multi_index_contains_overload(self, label):
     if not isinstance(self, MultiIndexType):
@@ -742,7 +658,7 @@ def pd_multi_index_contains_overload(self, label):
     def pd_multi_index_contains_impl(self, label):
 
         # build indexer_map (should already been built in index ctor?)
-        indexer_map = _build_index_map(self)
+        indexer_map = sdc_indexes_build_map_positions(self)
         res = label in indexer_map
         return res
 
@@ -915,53 +831,6 @@ def pd_multi_index_reindex_overload(self, target, method=None, level=None, limit
     return pd_multi_index_reindex_impl
 
 
-# TO-DO: seems like this can be removed when indexes have map_positions property
-@register_jitable
-def _appender_build_map(index1, index2):
-    res = {}
-    for i, val in enumerate(index1):
-        if val not in res:
-            res[val] = i
-
-    k, count = i, len(res)
-    while k < i + len(index2):
-        val = index2[k - i]
-        if val not in res:
-            res[val] = count
-            count += 1
-        k += 1
-
-    return res
-
-
-def _multi_index_append_level(A, codes_A, B, codes_B):
-    pass
-
-
-@sdc_overload(_multi_index_append_level)
-def _multi_index_append_overload(A, codes_A, B, codes_B):
-
-    def _multi_index_append_impl(A, codes_A, B, codes_B):
-
-        appender_map = _appender_build_map(A, B)
-        res_size = len(codes_A) + len(codes_B)
-        res_level = fix_df_index(
-            list(appender_map.keys())
-        )
-
-        res_codes = np.empty(res_size, dtype=np.int64)
-        A_size = len(codes_A)
-        for i in prange(res_size):
-            if i < A_size:
-                res_codes[i] = codes_A[i]
-            else:
-                res_codes[i] = appender_map[B[codes_B[i - A_size]]]
-
-        return (res_level, res_codes)
-
-    return _multi_index_append_impl
-
-
 @sdc_overload_method(MultiIndexType, 'append')
 def pd_multi_index_append_overload(self, other):
     if not isinstance(self, MultiIndexType):
@@ -992,115 +861,6 @@ def pd_multi_index_append_overload(self, other):
         )
 
     return pd_multi_index_append_impl
-
-
-def _multi_index_create_level(index_data, name):
-    pass
-
-
-@sdc_overload(_multi_index_create_level)
-def _multi_index_create_level_ovld(index_data, name):
-
-    def _multi_index_create_level_impl(index_data, name):
-        index = fix_df_index(index_data)
-        return sdc_indexes_rename(index, name)
-    return _multi_index_create_level_impl
-
-
-def _multi_index_create_levels_and_codes(level_data, codes_data, name):
-    pass
-
-
-@sdc_overload(_multi_index_create_levels_and_codes)
-def _multi_index_create_levels_and_codes_ovld(level_data, codes_data, name):
-
-    def _multi_index_create_levels_and_codes_impl(level_data, codes_data, name):
-        level_data_fixed = fix_df_index(level_data)
-        level = sdc_indexes_rename(level_data_fixed, name)
-        codes = fix_df_array(codes_data)
-
-        # to avoid additional overload make data verification checks inplace
-        # these checks repeat those in MultiIndex::_verify_integrity
-        if len(codes) and np.max(codes) >= len(level):
-            raise ValueError(
-                "On one of the levels code max >= length of level. "
-                "NOTE: this index is in an inconsistent state"
-            )
-        if len(codes) and np.min(codes) < -1:
-            raise ValueError(
-                "On one of the levels code value < -1")
-
-        # TO-DO: support is_unique for all indexes and use it here
-        indexer_map = _build_index_map(level)
-        if len(level) != len(indexer_map):
-            raise ValueError("Level values must be unique")
-
-        return (level, codes)
-
-    return _multi_index_create_levels_and_codes_impl
-
-
-@register_jitable
-def next_codes_info(level_info, cumprod_list):
-    _, codes = level_info
-    cumprod_list.append(cumprod_list[-1] * len(codes))
-    return codes, cumprod_list[-1]
-
-
-@register_jitable
-def next_codes_array(stats, res_size):
-    codes_pattern, factor = stats
-    span_i = res_size // factor                             # tiles whole array
-    repeat_i = res_size // (len(codes_pattern) * span_i)    # repeats each element
-    return np.array(list(np.repeat(codes_pattern, span_i)) * repeat_i)
-
-
-def factorize_level(level):
-    pass
-
-
-@sdc_overload(factorize_level)
-def factorize_level_ovld(level):
-
-    level_dtype = level.dtype
-
-    def factorize_level_impl(level):
-        unique_labels = List.empty_list(level_dtype)
-        res_size = len(level)
-        codes = np.empty(res_size, types.int64)
-        if not res_size:
-            return unique_labels, codes
-
-        indexer_map = Dict.empty(level_dtype, types.int64)
-        for i in range(res_size):
-            val = level[i]
-            _code = indexer_map.get(val, -1)
-            if _code == -1:
-                new_code = len(unique_labels)
-                indexer_map[val] = new_code
-                unique_labels.append(val)
-            else:
-                new_code = _code
-
-            codes[i] = new_code
-
-        return unique_labels, codes
-
-    return factorize_level_impl
-
-
-def _make_level_unique(index):
-    pass
-
-
-@sdc_overload(_make_level_unique)
-def _make_level_unique_ovld(index):
-
-    def _make_level_unique_impl(index):
-        indexer_map = _build_index_map(index)
-        return list(indexer_map.keys())
-
-    return _make_level_unique_impl
 
 
 @sdc_overload_method(MultiIndexTypeRef, 'from_product', prefer_literal=False)
@@ -1166,69 +926,6 @@ def pd_multi_index_from_product_overload(cls, iterables, sortorder=None, names=N
     return pd_multi_index_from_product_impl
 
 
-def _make_level_dict(index):
-    pass
-
-
-@sdc_overload(_make_level_dict)
-def _make_level_dict_ovld(index):
-
-    index_type = index
-
-    def _make_level_dict_impl(index):
-        return Dict.empty(index_type, types.int64)
-
-    return _make_level_dict_impl
-
-
-def _multi_index_get_new_code(level, val):
-
-    _code = level.get(val, -1)
-    if _code == -1:
-        res = len(level)
-        level[val] = res
-    else:
-        res = _code
-
-    return types.int64(res)
-
-
-def _multi_index_set_new_code(codes, new_code, i):
-    codes[i] = new_code
-
-
-@intrinsic
-def _multi_index_append_value(typingctx, val, levels, codes, idx):
-
-    nlevels = len(val)
-    if not (nlevels == len(levels) and nlevels == len(codes)):
-        assert True, f"Cannot append MultiIndex value to existing codes/levels.\n" \
-                     f"Given: val={val}, levels={levels}, codes={codes}"
-
-    def codegen(context, builder, sig, args):
-        index_val, levels_val, codes_val, idx_val = args
-
-        for i in range(nlevels):
-            label = builder.extract_value(index_val, i)
-            level_i = builder.extract_value(levels_val, i)
-            codes_i = builder.extract_value(codes_val, i)
-
-            new_code = context.compile_internal(
-                builder,
-                _multi_index_get_new_code,
-                signature(types.int64, levels[i], val[i]),
-                [level_i, label]
-            )
-            context.compile_internal(
-                builder,
-                _multi_index_set_new_code,
-                signature(types.none, codes[i], types.int64, idx),
-                [codes_i, new_code, idx_val]
-            )
-
-    return types.none(val, levels, codes, idx), codegen
-
-
 @sdc_overload_method(MultiIndexTypeRef, 'from_tuples', prefer_literal=False)
 def pd_multi_index_from_tuples_overload(cls, iterables):
     if cls.instance_type is not MultiIndexType:
@@ -1247,9 +944,10 @@ def pd_multi_index_from_tuples_overload(cls, iterables):
         if not index_size:
             raise TypeError("Cannot infer number of levels from empty list")
 
+        # use first value to infer types and allocate dicts for result multi index levels
         example_value = iterables[0]
         levels_dicts = sdc_tuple_map(
-            _make_level_dict,
+            _multi_index_alloc_level_dict,
             example_value
         )
         index_codes = sdc_tuple_map(
@@ -1258,9 +956,8 @@ def pd_multi_index_from_tuples_overload(cls, iterables):
             index_size
         )
 
-        for i in range(index_size):
-            val = iterables[i]
-            _multi_index_append_value(val, levels_dicts, index_codes, i)
+        for i, val in enumerate(iterables):
+            _multi_index_from_tuples_helper(val, levels_dicts, index_codes, i)
 
         index_levels = sdc_tuple_map(
             lambda x: list(x.keys()),
